==> 03 Scripts/01 Getting Familiar with customers dataframe.ipynb <==
# Summary of Notebook
    Full set: 206209
    Exported as pickle & saved as cust_ready_to_clean
    Memory: Reduced memory of original file
    Datatype:
        Changed the datatypes: Note user_id = int32
    Null: Null only in first_name column
    Duplicates:  No duplicates found
    Statistics:
        Age, Dependents, Income in reasonable range
    Column Notes:
        state looks reasonable
        gender looks reasonable
        living_situation may have issues. Interpretation may be sketcy: overlap possible with categories - divorced/widowed may also be single, married with parents and siblings may also be single

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_cust = pd.read_csv(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Original Data\customers.csv')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

#getting to know the dataframe
df_cust.shape

#seeing the type of data and columns
df_cust.info()

#seeing the columns and data
df_cust.head()

#seeing the end of the columns and data
df_cust.tail()

#renaming for consistency
df_cust.rename(columns={'First Name': 'first_name', 'Surnam': 'last_name', 'Gender':'gender', 'STATE': 'state', 'Age':'age', 'n_dependants':'dependents', 'fam_status':'living_situation'}, inplace=True)

#checking for change

df_cust.head()

#seeing info again
df_cust.info()

#changing datatype to reduce memory - changing objects to category
df_cust[['first_name', 'last_name','gender','state','living_situation']] = df_cust[['first_name','last_name','gender','state','living_situation']].astype('category')

#changing to date variable
df_cust['date_joined'] = df_cust['date_joined'].astype('datetime64')

#seeing what kind of interger type is best
df_cust.describe()

#changing datatype to reduce memory - datatype int16
df_cust[['age', 'dependents']] = df_cust[['age', 'dependents']].astype("int16")

#changing datatype to reduce memory - datatype int32
df_cust[['user_id', 'income']] = df_cust[['user_id', 'income']].astype("int32")

#checking dataframe
df_cust.info()

# Notes on memory change:  change happened

#seeing if any data is missing
df_cust.isnull().sum()

# Notes on null findings: first_name has missing data

#looking for duplicates
(df_cust.duplicated().value_counts())

# Notes on duplicates: no duplicates found

#checking to see if there is anything goofy with states
pivot = df_cust.pivot_table(index =['state'],
                       values =['income'],
                       aggfunc ='sum')
print(pivot)

# Notes on Pivot Column state: Quick glance, nothing out of the ordinary

#checking the gender column
pivot = df_cust.pivot_table(index =['gender'],
                       values =['income'],
                       aggfunc ='sum')
print(pivot)

# Notes on Pivot Column gender: Nothing out of the ordinary

#checking the living_situation column
pivot = df_cust.pivot_table(index =['living_situation'],
                       values =['income'],
                       aggfunc ='sum')
print(pivot)

# Notes on Pivot Column living situation: categories may be overlapping - divorced/widowed may also be single, married with parents and siblins may also be single

#exporting the file as a pickle
df_cust.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'cust_ready_to_clean.pkl'))
