==> 03 Scripts/01 Getting Familiar with departments dataframe.ipynb <==
# Summary of Notebook
    Full Set: 21
    Exported as pickle & saved as depart_transformed
    Transfomed the original dataframe from wide to long and cleaned up index
    Memory: kept it as it was
    Datatype: kept it as object
    Null: department_id 21 has department name as 'missing'
    Duplicates: no duplicates
    Column notes:  department_id 21 has missing department name

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_depart = pd.read_csv(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Original Data\departments.csv')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_depart = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'))

#getting to know the dataframe
df_depart.shape

#seeing the type of data and columns
df_depart.info()

#seeing the columns and data
df_depart.head()

#not an easy way to see the information: transposing the dataframe
df_depart.T

#renaming the new dataframe
df_depart_t = df_depart.T

df_depart_t.head()

#seeing the end of the columns and data
df_depart_t.tail()

#because the new database does not have an index - there was just a 0 as a header in the transposed
df_depart_t.reset_index()

 #getting rid of the 0 header by creating a new header and remember row 0 = 1st row
new_header = df_depart_t.iloc[0]

#checking to see the new output after changing the rule for the new header
new_header

 #create a new dataframe that only copies over rows beyond the first row
df_depart_t_new = df_depart_t[1:]

 #now sets the new header for the database
df_depart_t_new.columns = new_header

#checking that the updated database is correct
df_depart_t_new.head() 

df_depart_t_new.tail()

#seeing info again
df_depart_t_new.info()

#changing datatype to reduce memory - changing objects to category
df_depart_t_new['department'] = df_depart_t_new['department'].astype('category')

#changing datatype to reduce memory - changing objects to category
df_depart_t_new['department'] = df_depart_t_new['department'].astype('object')

#checking dataframe
df_depart_t_new.info()

# Notes on memory change:  kept it as it was

#seeing if any data is missing
df_depart_t_new.isnull().sum()

# Notes on null findings: nothing is null

#looking for duplicates
(df_depart_t_new.duplicated().value_counts())

# Notes on duplicates: no duplicates 

#exporting the file as a pickle
df_depart_t_new.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'depart_transformed.pkl'))
