==> 03 Scripts/01 Getting Familiar with orders dataframe.ipynb <==
# Summary of Notebook
    Full set:  3421093
    Exported as pickle & saved as orders_ready_to_clean
    Memory: Reduced memory of original file
    Datatype:
        user_id & order_id = int32
    Null:
        days_since_prior_order has 206,209 missing values
    Duplicates:  No duplicates
    Statistics:
        order_number ranges from 0-100
        day_of_week ranges from 0-6
        hour_of_day_ordered ranges from 0-23
        days since prior order ranges from 0 to 30

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_orders = pd.read_csv(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Original Data\orders.csv')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'))

#getting to know the dataframe
df_orders.shape

#seeing the type of data and columns
df_orders.info()

#seeing the columns and data
df_orders.head()

#seeing the end of the columns and data
df_orders.tail()

#renaming for consistency
df_orders.rename(columns={'order_dow': 'day_of_week_ordered', 'order_hour_of_day': 'hour_of_day_ordered'}, inplace=True)

#checking for chnage
df_orders.head()

#seeing info again
df_orders.info()

#seeing what kind of interger type is best
df_orders.describe()

#changing datatype to reduce memory - datatype int16
df_orders[['order_number', 'day_of_week_ordered','hour_of_day_ordered']] = df_orders[['order_number', 'day_of_week_ordered','hour_of_day_ordered']].astype("int16")

#changing datatype to reduce memory - datatype int32
df_orders[['order_id','user_id']] = df_orders[['order_id','user_id']].astype("int32")

#changing datatype to reduce memory - datatype category
df_orders['eval_set'] = df_orders['eval_set'].astype('category')

#checking dataframe
df_orders.info()

# Notes on memory change:  

#seeing if any data is missing
df_orders.isnull().sum()

# Notes on null findings: days_since_prior order has 206,209 missing values

#looking for duplicates
(df_orders.duplicated().value_counts())

# Notes on duplicates: no duplicates

#exporting the file as a pickle
df_orders.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_ready_to_clean.pkl'))
