==> 03 Scripts/01 Getting Familiar with orders_prior dataframe.ipynb <==
# Summary of Notebook
    Full Set: 32434489
    Exported as pickle & saved as orders_prior_ready_to_clean
    Memory: Reduced memory of original file
    Datatype:
        order_id and product_id = int32
    Null: No null
    Duplicates: No duplicates
    Statistics:
        add _to_cart order ranged from 1 to 145
        reordered ranged from 0 to 1

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_orders_prior = pd.read_csv(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Original Data\order_products_prior.csv')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products_prior.csv'))

#getting to know the dataframe
df_orders_prior.shape

#seeing the type of data and columns
df_orders_prior.info()

#seeing the columns and data
df_orders_prior.head()

#seeing the end of the columns and data
df_orders_prior.tail()

#seeing info again
df_orders_prior.info()

#seeing what kind of interger type is best
df_orders_prior.describe()

df_orders_prior[['add_to_cart_order', 'reordered']] = df_orders_prior[['add_to_cart_order', 'reordered']].astype("int16")

#changing datatype to reduce memory - datatype int32
df_orders_prior[['order_id', 'product_id']] = df_orders_prior[['order_id', 'product_id']].astype("int32")

#checking dataframe
df_orders_prior.info()

# Notes on memory change: Reduced memory usage 

#seeing if any data is missing
df_orders_prior.isnull().sum()

# Notes on null findings: No null information 

#looking for duplicates
df_orders_prior.duplicated().value_counts()

# Notes on duplicates: No duplicates

#exporting the file as a pickle
df_orders_prior.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_prior_ready_to_clean.pkl'))
