==> 03 Scripts/01 Getting Familiar with products dataframe.ipynb <==
# Summary of Notebook
    Full Set: 49688
    Exported as pickle & saved as products_ready_to_clean
    Memory: Reduced memory of original file
    Datatype:
        product_id = int32
        department_id = int16
    Null:
        product_name has 16 products without a name with a high price range (so more costly items)
    Duplicates: dropped 5 duplicates
    Statistics:
       product_id has a range of 1 to 49688
       aisle_id has a range of 1 to 134
       price has a range of 1 to 99999 

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_products = pd.read_csv(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Original Data\products.csv')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_products = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'))

#getting to know the dataframe
df_products.shape

#seeing the type of data and columns
df_products.info()

#seeing the columns and data
df_products.head()

#seeing the end of the columns and data
df_products.tail()

#renaming for consistency
df_products.rename(columns={'prices': 'price'}, inplace=True)

#checking for chnage
df_products.head()

#seeing info again
df_products.info()

#changing datatype to reduce memory - changing objects to category
df_products['product_name'] = df_products['product_name'].astype('category')

#seeing what kind of interger type is best
df_products.describe()

#changing datatype to reduce memory - datatype int16
df_products[['aisle_id', 'department_id']] = df_products[['aisle_id', 'department_id']].astype("int16")

#changing datatype back to int32 because I messed up
df_products ['product_id'] = df_products ['product_id'].astype("int32")

#checking dataframe
df_products.info()

# Notes on memory change:  no real change

#seeing if any data is missing
df_products.isnull().sum()

# Notes on null findings: there are 16 products without a name

#printing more information on the products with missing names
nan_values = df_products[df_products['product_name'].isna()]

print (nan_values)

#looking for duplicates
(df_products.duplicated().value_counts())

#curious to see what was duplicated
df2 = df_products[df_products.duplicated()]
print(df2)

df_products.drop_duplicates()

# Notes on duplicates: 5 duplicates found. Dropped them

#need to save the new dataframe without duplicates
df_products = df_products.drop_duplicates()

#checking the result
df_products.describe()

#exporting the file as a pickle
df_products.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'products_ready_to_clean.pkl'))
