==> 03 Scripts/02 Active Customer EDA.ipynb <==
# Summary of Notebook
    This is the active customers dataframe with all the information within the dataframe: These customers had more than a max order >5
    It pulled through with the correct number of rows
    Double checked the memory: 2.6 GB
    No null values
    No duplicate rows
    I was unable to run EDA packages because my computer didn't have enough memory, so I used describe to see how this dataset is like with only the active customers
        notes
            day_of_week_ordered ranges from 0-6
            hour_of_day_ordered ranges from 0 -23
            days_since_prior_order ranges from 0 - 30
            add_to_cart_order ranges from 1 to 145 (I had checked this previously and yes, some customers had large orders)
            reordered ranges from 0-1 (this is correct - those are the only options)
            aisle_id ranges from 1 to 134
            department_id ranges from 1-21
            price ranges from 1 to 25
            age ranges from 18 to 81
            dependents ranges from 0 to 3
            income ranges from 25903 to 593901
            max_order ranges from 6 to 99
            total_spent ranges from 6 to 25792.9
            median_days_prior_ord ranges from 0 to 30
    

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_active = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\04 Analysis\active_all_plus_regions.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_active = pd.read_pickle(os.path.join(path, '04 Analysis', 'active_all_plus_regions.pkl'))

df_active.info()

#no data should be missing
df_active.isnull().sum()

#should be no duplicates
(df_active.duplicated().value_counts())

df_active.describe()

# notes
    day_of_week_ordered ranges from 0-6
    hour_of_day_ordered ranges from 0 -23
    days_since_prior_order ranges from 0 - 30
    add_to_cart_order ranges from 1 to 145 (I had checked this previously and yes, some customers had large orders)
    reordered ranges from 0-1 (this is correct - those are the only options)
    aisle_id ranges from 1 to 134
    department_id ranges from 1-21
    price ranges from 1 to 25
    age ranges from 18 to 81
    dependents ranges from 0 to 3
    income ranges from 25903 to 593901
    max_order ranges from 6 to 99
    total_spent ranges from 6 to 25792.9
    median_days_prior_ord ranges from 0 to 30
