==> 03 Scripts/03 Cleaning Customer Dataframe.ipynb <==
# Summary of Notebook
    Exported datafram 206209 rows
    Exported as pickle
    Saved as: cust_cleaned
    Dropped first and last name columns

# Summary of Previous Notebook
    Exported dataframe had 206209 rows
    Exported as pickle named: cust_ready_to_clean
    user_id = datatype int32
    Used dataprep to analyze customer dataframe
    Privacy issues: has first and last name
    First name has missing data 11,259 or 5.5% of the column
    No duplicates
    Low correlation between age and income: .4
    Gender is distributed 50/50
    Each state has a 4055 count
    Age ranges from 18 to 81
    Dependents ranges from 0 to 3 with a 25% distribution of each number of dependents
    Living situation:70.2% indicated married status
    Living situation may have overlapping info: living with parents, widowed/divorced and single may overlap depending on the person's interpretation
    Income ranged from $25,903 to $593,901
    Renamed columns
    Changed datatype to reduce memory

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv customers dataframe
df_cust = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\cust_ready_to_clean.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'cust_ready_to_clean.pkl'))

#checking to ensure all data in dataframe
df_cust.info()

#dataframe imported correctly

#need to ensure privacy - first_name and last_name should not be included for analysis
df_cust = df_cust.drop('first_name', axis=1)

df_cust = df_cust.drop('last_name', axis = 1)

df_cust.info()

#exporting the file as a pickle
df_cust.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'cust_cleaned.pkl'))
