==> 03 Scripts/03 Cleaning Orders Dataframe.ipynb <==
# Summary of Notebook
    used a crosstab command to better understand the missing days_since_prior_orders: result: first orders all had 0 days since prior order. It seems to make sense that the days_since_prior_order having missing data = first time orders.
        Exported the one-time customer as pickle with 206209 rows
        Saved as one_time_purchase.pkl
    order_id and user_id both = int32 datatype
    Dropped the missing values from original dataframe
    Dropped eval_set column from dataframe
    Exported as pickle with 3214874 rows
    Saved as orders_cleaned.pkl
    

# Summary of Previous Notebook
    Dataframe exported with 3421083 rows
    Dataframe exported as pickle
    Saved as: orders_ready_to_clean.pkl
    Renamed variables for clarity
    order_id & user_id = int32
    Changed datatype to conserve memory
    Used dataprep to analyze orders dataframe
    Missing data: days_since_prior_order is missing 206,209 times (6.03% missing for the variable)
    No duplicate rows
    user_id: approximately 6% is unique. These unique matches the missing which = one time order
    eval_set = prior, train, test. 94% = prior
    order_number ranges from 1 to 100
    day_of_week_ordered seen in dataprep as a categorical versus int16, so had to view that differently
    hour_of_day_ordered ranges from 0 to 23
    days_since_prior_order ranges from 0 to 30; 0 & .6 days totals to 6.25% of the variable count
    no correlations seen
    using describe: day_of_week_ordered ranges from 0 to 6
    Included sweetviz as an output for visual as well

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import csv orders dataframe
df_orders = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_ready_to_clean.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_ready_to_clean.pkl'))

#confirming full dataframe imported
df_orders.info()

#imported dataframe matches

#days_since_prior_order is missing 206,209 times - figuring out why. I assume it is due to customers only placing one order in the system
crosstab = pd.crosstab(df_orders['days_since_prior_order'], df_orders['order_number'], dropna = False)

crosstab.to_clipboard()

#when I looked at the data in Excel, choosing "filter" for the column of 1 order (which = first order), every row had 0 days since prior order, so the missing values are for one time orders

# learning the one time only customers
null_data = df_orders[df_orders.isnull().any(axis=1)]

#checking the dataframe head
null_data.head()

#double checking that it has the correct number of rows
null_data.info()

#exporting the one time only purchasers in case a comparison is needed
#exporting the file as a pickle
null_data.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'one_time_purchase.pkl'))

#dropping the rows with missing values for days_since_prior_order
df_orders = df_orders.dropna()

#dropping eval_set
df_orders = df_orders.drop('eval_set', 1)

df_orders.info()

#exporting the file as a pickle
df_orders.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned.pkl'))
