==> 03 Scripts/03 Cleaning Orders-Prior Dataframe.ipynb <==
# Summary of Notebook
    Checked the very large orders with >100 add_add_to_cart_order: seemed legit
    Dataframe 32434489
    Exported as pickle
    Saved as orders_prior_clean.pkl

# Summary of Previous Notebook
    Dataframe rows: 32434489
    Exported as pickle & saved as orders_prior_ready_to_clean
    Changed the datatypes to conserve memory
    order_id & product_id datatype = int32
    No duplicates
    No missing values
    10% of order_id are distinct: one item orders
    product_id: range 1 to 49,677
    add_to_cart_order: range 1 to 145; there are orders with 131 to 145 add to card order - kind of suspicious
    reordered: range 0-1; 59% of the time an item was reordered

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_prior_ready_to_clean dataframe
df_orders_prior = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_prior_ready_to_clean.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders_prior = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_prior_ready_to_clean.pkl'))

#checking import
df_orders_prior.info()

#import matches

#trying to figure out the high add_to_cart_order rows
df_orders_prior.groupby(['order_id', 'product_id'])[['add_to_cart_order']].agg('count')

df_orders_prior.value_counts()

df_orders_prior.value_counts(['add_to_cart_order'])

limit = 100

column_name = 'add_to_cart_order'

column = df_orders_prior['add_to_cart_order']

count = column[column > limit].count()

print('Count of add_to_cart_order > 100: ', count)

df_orders_prior['add_to_cart_order'].value_counts().sort_index(ascending = True).sort_values(ascending = True)

df_orders_prior.loc[(df_orders_prior['add_to_cart_order'] > 100)]

df_very_large = df_orders_prior.loc[(df_orders_prior['add_to_cart_order'] > 100)]

df_very_large.to_clipboard()

#I looked at the exported file and I can't believe there would be huge orders like that, but it appeared legit.

df_orders_prior.info()

#exporting the file as a pickle
df_orders_prior.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_prior_clean.pkl'))
