==> 03 Scripts/03 Cleaning Products Dataframe.ipynb <==
# Summary of Notebook
    Dataframe: 49677
    Exported as pickle
    Saved as:  products_cleaned.pkl
    Learned the price mode. Used price mode to fix items with erroneous prices
    Changed the price of milk from 99999 to 2.5
    Changed the price of cottage cheese from 149000 to 2.5
    price range now 1-25
    

# Summary of Previous Notebook
    Dataframe: 49693 rows
    Exported as pickle
    Saved as: products_ready_to_clean
    Changed datatype
        product_id = int32
    product_name is missing in 16 rows
    product_id range: 1 to 49688
    aisle_id range: 1 to 134
    department_id range: 1 to 21
        highest count in department_id: 11, 19 and 13
    prices: $1 to $99,999

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import pickle products dataframe
df_products = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\products_ready_to_clean.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_products = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'products_ready_to_clean.pkl'))

df_products.info()

#imported with correct number of rows

#checking values
df_products.describe()

#dropping rows with no product name
df_products = df_products.dropna()

#checking change
df_products.info()

#seeing if I can figure out the high price
df_products.loc[df_products['price'] == 99999]

#getting the mode for the column
df_products.mode()['price'][0]

#changing the 99999 price
df_products.at[33666,'price']=2.5

df_products.describe()

#seeing if I can figure out the high price
df_products.loc[df_products['price'] == 14900]

#changing the 14900 price
df_products.at[21554,'price']=2.5

df_products.describe()

#exporting the file as a pickle
df_products.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'products_cleaned.pkl'))
