==> 03 Scripts/04 All orders combined.ipynb <==
# Summary of Notebook
    merged orders_cleaned with orders_prior_clean 
    performed an inner merge so only the order_id within both sets was merged together
    unable to compare what was lost from each dataframe because I had a memory error when attempting an outer join
    merged dataframe = 30356421
    exported dataframe as a pickle
    saved as: all_orders.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_cleaned
df_orders = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_cleaned.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_cleaned.pkl'))

#checking the rows
df_orders.info()

#import orders__prior_clean
df_orders__prior = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_prior_clean.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders_prior = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_prior_clean.pkl'))

#checking the rows
df_orders_prior.info()

#merging the dataframes
df_merged_orders = df_orders.merge(df_orders_prior, on = ('order_id'), indicator = True)

df_merged_orders.head()

# checking the rows merged on order_id
df_merged_orders['_merge'].value_counts()

df_merged_orders.info()

#please note: I could not do an outer merge to check because my computer did not have enough memory

#seeing if any data is missing
df_merged_orders.isnull().sum()

#need to drop the _merge column
df_merged_orders = df_merged_orders.drop('_merge', axis=1)

df_merged_orders.head()

df_merged_orders.info()

#exporting all orders as a pickle
df_merged_orders.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'all_orders.pkl'))
