==> 03 Scripts/04 Orders and products combined.ipynb <==
# Summary of Notebook
    inner join orders on products to reduce null values
    after merging orders with products, the merged file is too large
    searched to understand why.... sleuthing indicated that there may have been duplicate product_ids
    7 duplicate IDs were found: I deleted the ones that were truly duplicates: fiber 4 gm gummy, ranger ipa, black house coffee,  gluten free peanut butter, body wash
    I could change a duplicate product_id to match the index: Revive Zero Vitamin Water now product_id 6799
    There was a duplicate_id for a deodorant and sharp cheddar cheese. After analysis, the deodorant was never purchased. I dropped the row containing the never purchased deodorant
    
    exported file: 30330084
    exported as pickle
    saved as: orders_products_combined.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import all_orders
df_orders = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\all_orders.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'all_orders.pkl'))

df_orders.info()

#import products_cleaned
df_products = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\products_cleaned.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_products = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'products_cleaned.pkl'))

df_products.info()

#seeing if any data is missing
df_products.isnull().sum()

#seeing if any data is missing
df_orders.isnull().sum()

# doing a inner join: I want all orders to have a product information and no nulls
df_combined = df_orders.merge(df_products, on = ('product_id'), indicator = True)

# checking the rows merged on product_id
df_combined['_merge'].value_counts()

df_combined.shape

df_orders.shape

#this isn't making sense. The order shape and the combined shape should be the same

#trying to see if there are duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

#dropping the duplicate rows
df_products = df_products.drop(labels=[462,18459,26810,35309,35495], axis=0)

#checking to see change in duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

# doing a inner join: I want all orders to have a product information and no nulls
df_combined = df_orders.merge(df_products, on = ('product_id'), indicator = True)

df_combined.shape

#trying to see if the products are listed twice or just need a new product_id
found = df_products[df_products['product_name'].str.contains('Revive Zero Vitamin Water')]
print(found.count())

#trying to see if the products are listed twice or just need a new product_id
found = df_products[df_products['product_name'].str.contains('Sprouted Quinoa Flakes Baby Cereal')]
print(found.count())

#trying to see if the products are listed twice or just need a new product_id
found = df_products[df_products['product_name'].str.contains('Clinical Advanced Solid Ultimate Fresh')]
print(found.count())

#trying to see if the products are listed twice or just need a new product_id
found = df_products[df_products['product_name'].str.contains('Cheese Shredded Sharp Cheddar Reduced Fat 2%')]
print(found.count())

#each item is only in the product_name column once so the name is not duplicated. Based on the pattern of index number and product_id, I am going to change the product_id to match the index number

#checking to see change in duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

#attempting to change the product_id
df_products.at[6799, 'product_id']

#attempting to change the product_id
df_products.at[6799, 'product_id']=6799

#checking to see change in duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

#attempting to change the product_id
df_products.at[26521, 'product_id']=26521

#checking to see change in duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

# doing a inner join: I want all orders to have a product information and no nulls
df_combined = df_orders.merge(df_products, on = ('product_id'), indicator = True)

df_combined.shape

df_orders.shape

#this isn't working If I were on a team, I'd just ask for the product_id

#attempting to change the product_id
df_products.at[26521, 'product_id']=26520

#checking to see change in duplicated product_id
df_products[df_products.duplicated(['product_id'], keep=False)]

# doing a inner join: I want all orders to have a product information and no nulls
df_combined = df_orders.merge(df_products, on = ('product_id'), indicator = True)

df_combined.shape

#counting how many times the product_name clinincal advances solid ultimate happens
(df_combined['product_name']=='Clinical Advanced Solit Ultimate Fresh').sum()

#counting how many times the product_name cheese shredded sharp cheddar reduced fat ultimate happens
(df_combined['product_name']=='Cheese Shredded Sharp Cheddar Reduced Fat 2%').sum()

#based on this analysis, I can drop the row named as index 26520

#dropping the duplicate row
df_products = df_products.drop(labels=[26520], axis=0)

# doing a inner join: I want all orders to have a product information and no nulls
df_combined = df_orders.merge(df_products, on = ('product_id'), indicator = True)

df_combined.shape

df_orders.shape

#seeing if any data is missing
df_combined.isnull().sum()

#exporting merged file as a pickle
df_combined.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
