==> 03 Scripts/04 Orders products & customers combined.ipynb <==
# Summary of Notebook
    left join orders_products_combined with cust_cleaned
    Data: 30330084 rows
    exported as pickle
    saved as orders_products_customers_combined.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import all_orders
df_orders = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_orders.info()

#import cust_cleaned dataframe
df_cust = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\cust_cleaned.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'cust_cleaned.pkl'))

df_cust.info()

#need to drop the _merge column
df_orders = df_orders.drop('_merge', axis=1)

#merging the dataframes: left merge so each order has customer info
df_merged = df_orders.merge(df_cust, on = ('user_id'), how = 'left',indicator = True)

df_merged.info()

#need to drop the _merge column
df_merged = df_merged.drop('_merge', axis=1)

df_merged.head()

#exporting all orders as a pickle
df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_customers_combined.pkl'))
