==> 03 Scripts/05 Deriving New Variable Active Customers.ipynb <==
# Summary of Notebook
    Added a new column: order_amount_paid
    order_amount_paid range: 1 to 1092.20
    Exported with 28669841 rows
    Exported as pickle
    Exported to 04 Analysis
    Saved as active_all_plus_regions_order_paid.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import pickle active customers dataframe
df_active = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\04 Analysis\active_all_plus_regions.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_active = pd.read_pickle(os.path.join(path, '04 Analysis', 'active_all_plus_regions.pkl'))

#checking import
df_active.info()

#learning the total amount for each order
#learning the maximum number of orders for each customer
df_active['order_paid_amount'] = df_active.groupby(['order_id'])['price'].transform(np.sum)

df_active.head()

df_active.describe()

#total paid for an order ranged from 1 to 1092.20

df_active.info()

#exporting the new dataframe as a pickle
df_active.to_pickle(os.path.join(path, '04 Analysis','active_all_plus_regions_order_paid.pkl'))
