==> 03 Scripts/05 Deriving New Variables - customers.ipynb <==
# Summary of Notebook
    defined loyalty_flag based on maximum number of orders
    defined spender_flag based on total spent
    defined frequency_flag based on median_days_prior_ord
    dataframe has 30330084 rows
    exported as pickle
    saved as '04 Analysis','cust_loyalty_spender_frequency.pkl'

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_products_customers_combined
df_all = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\04 Analysis\all_busy_day_hour.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_all = pd.read_pickle(os.path.join(path, '04 Analysis', 'all_busy_day_hour.pkl'))

df_all.info()

#learning the maximum number of orders for each customer
df_all['max_order'] = df_all.groupby(['user_id'])['order_number'].transform(np.max)

df_all.head()

#defining loyal customers based on the number of max orders
df_all.loc[df_all['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

#defining average customers based on the number of max orders
df_all.loc[(df_all['max_order'] <= 40) & (df_all['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

#defining new customers based on the number of max orders
df_all.loc[df_all['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

#seeing the number of customers in each category
df_all['loyalty_flag'].value_counts(dropna=False)

df_all.head()

#learning the total spent for each customer
df_all['total_spent'] = df_all.groupby(['user_id'])['price'].transform(np.sum)

df_all['total_spent'].describe()

#defining high spending customers based on total spent
df_all.loc[df_all['total_spent'] >= 4663, 'spender_flag'] = 'High spender'

#defining low spending customers based on total spent
df_all.loc[df_all['total_spent'] < 1106, 'spender_flag'] = 'Low spender'

#defining average spending customers based on total spent
df_all.loc[(df_all['total_spent'] <4663) & (df_all['total_spent'] >=1106), 'spender_flag'] = 'Average spender'

#seeing number of customers in spending category
df_all['spender_flag'].value_counts(dropna=False)

df_all.head()

#so creating a group based on the customer loyalty
df_all.groupby('loyalty_flag')

#to learn the type of spender based on customer loyalty
df_all.groupby('loyalty_flag').agg({'total_spent': ['mean']})

#looking at the frequency of spending
df_all.groupby('user_id')

#median of days_since_prior_order
df_all['days_since_prior_order'].median()

#learning the median of days_since_prior_order for each customer
df_all['median_days_prior_ord'] = df_all.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

#checking for change
df_all[['user_id', 'loyalty_flag', 'spender_flag',  'median_days_prior_ord']].head()

#defining low frequency customers based on median_days_prior_ord
df_all.loc[df_all['median_days_prior_ord'] >= 20, 'frequency_flag'] = 'Non-frequent customer'

#defining average frequency customers based on median_days_prior_ord
df_all.loc[(df_all['median_days_prior_ord'] < 20) & (df_all['median_days_prior_ord'] > 10), 'frequency_flag'] = 'Regular customer'

#defining high frequency customers based on median_days_prior_ord
df_all.loc[df_all['median_days_prior_ord'] <= 10, 'frequency_flag'] = 'Frequent customer'

#seeing number of customers in frequency flag
df_all['frequency_flag'].value_counts(dropna=False)

df_all.head()

output = pd.pivot_table(data=df_all, 
                        index=['loyalty_flag','spender_flag','frequency_flag'], 
                        values='total_spent',
                        aggfunc='mean')
output

#checking dataframe
df_all.info()

#exporting the new dataframe as a pickle
df_all.to_pickle(os.path.join(path, '04 Analysis','cust_loyalty_spender_frequency.pkl'))
