==> 03 Scripts/05 Deriving New Variables - product price range, busiest day, busiest hour.ipynb <==
# Summary of Notebook
    qualified products as high-range, mid-range or low-range based on price
    qualified busiest day of week as busiest day, regularly busy or least busy based on the count for each day of week
    qualified hour of day as most orders, average orders, fewest orders based on the count for each hour of the day
    Exported rows: 30330084
    Saved as pickle
    Saved in analysis
    Saved as all_busy_day_hour.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_products_customers_combined
df_all = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\02 Data\Prepared Data\orders_products_customers_combined.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_customers_combined.pkl'))

df_all.head()

df_all.info()

#seeing if any data is missing
df_all.isnull().sum()

#only going to work with a subset of the data and look at the first million rows
df = df_all[:1000000]

#using .loc to create price categories
df.loc[df['price'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['price'] <= 15) & (df['price'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['price'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

df_all.loc[df_all['price'] >15,'price_range_loc'] = 'High-range product'

df_all.loc[(df_all['price'] <=15) & (df_all['price'] >5), 'price_range_loc'] = 'Mid-range product'

df_all.loc[df_all['price'] <=5,'price_range_loc'] = 'Low-range product'

df_all['price_range_loc'].value_counts(dropna = False)

#which day of week has highest frequency of orders?
df_all['day_of_week_ordered'].value_counts(dropna = False)

#now adding meaning to the data. Defining the busy level for day of week
result = []

for value in df_all["day_of_week_ordered"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 1:
    result.append("Busiest day")
  elif value == 3:
    result.append("Least busy")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

#for everything to make sense, need to add the result to the df
df_all['busiest_day'] = result

#looking at the frequency of the levels of categories of busy and number matches original day of week data
df_all['busiest_day'].value_counts(dropna = False)

df_all.head()

#which hour of day has highest frequency of orders?
df_all['hour_of_day_ordered'].value_counts(dropna = False)

#now categorizing the busiest order times
result = []

for value in df_all["hour_of_day_ordered"]:
  if value == 10:
    result.append("Most Orders")
  elif value == 11:
    result.append("Most Orders")
  elif value == 14:
    result.append("Most Orders")
  elif value == 15:
    result.append("Most Orders")
  elif value == 13:
    result.append("Most Orders")
  elif value == 12:
    result.append("Most Orders")
  elif value == 5:
    result.append("Fewest Orders")
  elif value == 2:
    result.append("Fewest Orders")
  elif value == 4:
    result.append("Fewest Orders")
  elif value == 3:
    result.append("Fewest Orders")
  else:
    result.append("Average Orders")

result

#for everything to make sense, need to add the result to the df
df_all['busiest_period_of_day'] = result

df_all.head()

#looking at the frequency of the levels of categories of busy and number matches original hour of day order placed data
df_all['busiest_period_of_day'].value_counts(dropna = False)

df_all.info()

#exporting the new dataframe as a pickle
df_all.to_pickle(os.path.join(path, '04 Analysis', 'all_busy_day_hour.pkl'))
