==> 03 Scripts/05 Deriving New Variables - region of country.ipynb <==
# Summary of Notebook
    created regions based on https://simple.wikipedia.org/wiki/List_of_regions_of_the_United_States
    created a region column based on states
    was able to group by region and spender_flag to see if differences in spending in region
    Exported 30330084 rows
    Exported to 04 Analysis
    Exported as pickle
    Saved as all_plus_regions.pkl

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_products_customers_combined
df_all = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\04 Analysis\cust_loyalty_spender_frequency.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_all = pd.read_pickle(os.path.join(path, '04 Analysis', 'cust_loyalty_spender_frequency.pkl'))

#checking import
df_all.shape

df_all.head()

df_all.info()

#creating regions
regions = {
    'Maine':'Northeast',
    'New Hampshire': 'Northeast',
    'Vermont': 'Northeast',
    'Massachusetts': 'Northeast',
    'Rhode Island':'Northeast',
    'Connecticut':'Northeast',
    'New York':'Northeast',
    'Pennsylvania': 'Northeast',
    'New Jersey': 'Northeast',
    'Wisconsin': 'Midwest',
    'Michigan': 'Midwest',
    'Illinois':'Midwest',
    'Indiana':'Midwest',
    'Ohio': 'Midwest',
    'North Dakota':'Midwest',
    'South Dakota': 'Midwest',
    'Nebraska': 'Midwest',
    'Kansas': 'Midwest',
    'Minnesota': 'Midwest',
    'Iowa':'Midwest',
    'Missouri':'Midwest',
    'Delaware':'South',
    'Maryland': 'South',
    'District of Columbia': 'South',
    'Virginia':'South',
    'West Virginia': 'South',
    'North Carolina': 'South',
    'South Carolina': 'South',
    'Georgia': 'South',
    'Florida': 'South',
    'Kentucky':'South',
    'Tennessee': 'South',
    'Mississippi': 'South',
    'Alabama': 'South',
    'Oklahoma': 'South',
    'Texas': 'South',
    'Arkansas': 'South',
    'Louisiana': 'South',
    'Idaho': 'West',
    'Montana': 'West',
    'Wyoming': 'West',
    'Nevada': 'West',
    'Utah': 'West',
    'Colorado': 'West',
    'Arizona':'West',
    'New Mexico': 'West',
    'Alaska': 'West',
    'Washington': 'West',
    'Oregon': 'West',
    'California': 'West',
    'Hawaii': 'West'    
}

#trying to create the regions onto the dataframe
def get_region(state):
    return regions[state]

df_all['region'] = df_all['state'].apply(get_region)

df_all.head()

#seeing number of customers in spending region
df_all['region'].value_counts(dropna=False)

df_all.groupby(["region", "spender_flag"]).size()

df_all.shape

#exporting the new dataframe as a pickle
df_all.to_pickle(os.path.join(path, '04 Analysis', 'all_plus_regions.pkl'))
