==> 03 Scripts/07 Created New Dataframes - customers with less than 5 orders and customers with more than 5 orders.ipynb <==
# Summary of Notebook
    Based on definition of active: located all the customers with =< 5 orders
    Exported 1660243 rows
    Exported to 04 Analysis
    Exported as pickle
    Saved as low_all_plus_regions.pkl
    Based on definition of active: located all customers with >5 orders
    Exported 28669841 rows
    Exported to 04 Analysis
    Exported as picklle
    Saved as active_all_plus_regions

#importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import datetime as dt

#import orders_products_customers_combined
df_all = pd.read_pickle(r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis\04 Analysis\all_plus_regions.pkl')
path = r'C:\Users\Selena\Desktop\3-22 Instacart Basket Analysis'
df_all = pd.read_pickle(os.path.join(path, '04 Analysis', 'all_plus_regions.pkl'))

df_all.info()

df_all.loc[df_all['max_order'] <= 5]

df_low = df_all.loc[df_all['max_order'] <= 5]

df_low.info()

#exporting the new dataframe as a pickle
df_low.to_pickle(os.path.join(path, '04 Analysis', 'low_all_plus_regions.pkl'))

#finding active customers
df_all.loc[df_all['max_order'] > 5]

#creating new dataframe
df_high = df_all.loc[df_all['max_order'] > 5]

df_high.info()

df_high.shape

#exporting the new dataframe as a pickle
df_high.to_pickle(os.path.join(path, '04 Analysis', 'active_all_plus_regions.pkl'))
